requirements.txt
langchain==0.3.22
langchain-core==0.3.49
langchain-community==0.3.20
langchain-openai==0.3.11
langgraph==0.3.23
markdownify==1.1.0
readability-lxml==0.8.1
pyhtml2md==1.6.6
lxml_html_clean==0.4.1
python-docx==1.1.2
pypdf==5.4.0
unstructured==0.17.2
pathlib==1.0.1
duckduckgo_search==2025.4.1
